3. What one has not experienced, one will never understand in print

+++ main.h
@@ -4,7 +4,7 @@
 #include <stdio.h>
 int _printf(const char *format, ...);
 int _print_string(int len, char *p);
-int _print_positif_int(int len, int n);
+int _print_positif_int(int len, long int n);
 int _print_negatif_int(int len, int n);
 int _print_NULL(int len);
 int _print_char(int len, const char *p);
